requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
aiosqlite==0.17.0
alembic==1.7.5
anyio==3.4.0
asgiref==3.4.1
async-timeout==4.0.2
asyncpg==0.25.0
atomicwrites==1.4.0
attrs==21.4.0
autobahn==22.5.1
Automat==20.2.0
base58==2.1.0
bcrypt==3.2.0
binance-connector==1.13.0
bitarray==1.2.2
bitcoin-utils-fork-minimal==0.4.11.4
bitcoinlib==0.6.4
block-io==2.0.5
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
constantly==15.1.0
cryptography==36.0.1
cytoolz==0.11.2
dateparser==1.0.0
distlib==0.3.4
dnspython==2.1.0
dydx-v3-python==1.6.0
ecdsa==0.16.0
email-validator==1.1.3
eth-abi==2.1.1
eth-account==0.5.7
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
fastapi==0.70.1
filelock==3.7.1
frozenlist==1.3.0
greenlet==1.1.2
h11==0.12.0
hexbytes==0.2.2
httptools==0.2.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==0.23
incremental==21.3.0
ipfshttpclient==0.8.0a2
iso8601==1.0.2
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.5.1
lru-dict==1.1.7
luno-python==0.0.9
Mako==1.1.6
MarkupSafe==2.0.1
more-itertools==8.13.0
mpmath==1.0.0
multiaddr==0.0.9
multidict==6.0.2
netaddr==0.8.0
numpy==1.22.3
orjson==3.6.5
packaging==21.3
parsimonious==0.8.1
passlib==1.7.4
paxful-api==1.0.1
Pillow==9.1.0
platformdirs==2.5.2
pluggy==0.13.1
protobuf==3.20.1
psycopg2==2.9.3
py==1.11.0
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.14.1
pydantic==1.9.0
PyJWT==2.3.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pypika-tortoise==0.1.4
pyrsistent==0.18.1
pytest==4.6.11
python-binance==1.0.16
python-dateutil==2.8.2
python-dotenv==0.19.2
python-http-client==3.3.7
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
# pywin32==304
PyYAML==5.4.1
regex==2022.6.2
requests==2.26.0
requests-mock==1.6.0
rlp==2.0.1
rsa==4.8
scrypt==0.8.20
sendgrid==6.9.7
service-identity==21.1.0
simplejson==3.17.6
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.29
starkbank-ecdsa==2.0.3
starlette==0.16.0
sympy==1.6
toml==0.10.2
toolz==0.11.2
tortoise-orm==0.19.0
tox==3.13.2
twilio==7.9.2
Twisted==22.4.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.0.1
tzdata==2022.1
tzlocal==4.2
ujson==4.3.0
urllib3==1.26.7
uvicorn==0.15.0
varint==1.0.2
virtualenv==20.14.1
watchgod==0.7
wcwidth==0.2.5
web3==5.29.0
websockets==9.1
yarl==1.7.2
zipp==3.8.0
zope.interface==5.4.0
